+++ examples/next/app/[[...slug]]/static-content-renderer.tsx
@@ -1,6 +1,7 @@
 import { StaticContent } from "@simple-builder/react";
 
 import "./builder-registery";
+import "@simple-builder/react/styles";
 
 type StaticContentRenderer = React.ComponentProps<typeof StaticContent>;