+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @simple-builder/react
 
+## 1.0.21
+
+### Patch Changes
+
+- Added option to hide component
+- Updated dependencies
+  - @simple-builder/server@1.0.21
+
 ## 1.0.20
 
 ### Patch Changes